apps/requirements.txt
aiohttp==3.7.4.post0
asgiref==3.5.2
async-timeout==3.0.1
attrs==21.4.0
black==22.3.0
certifi==2022.5.18.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
cycler==0.11.0
discord==1.7.3
discord.py==1.7.3
Django==3.2.13
django-model-utils==4.2.0
fonttools==4.33.3
idna==3.3
importlib-metadata==4.11.4
jsonfield==3.1.0
kiwisolver==1.4.2
matplotlib==3.5.2
multidict==6.0.2
mypy-extensions==0.4.3
numpy==1.21.6
packaging==21.3
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
psycopg2==2.9.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typed-ast==1.5.4
typing-extensions==4.2.0
urllib3==1.26.9
yarl==1.7.2
zipp==3.8.0
